6 Kyu this took so fucking long. I needed to chatgpt for a second. I didn't know how to create the array I wanted and put values in it which was my main problem. also fucked up the median calculation.
not looking at the top code rn, but it's:

package solution

object Stat {

    fun stat(s: String): String {
        fun String.toSecs() = split("|").map { it.toInt() }.let { it[0] * 3600 + it[1] * 60 + it[2] }
        fun Int.toHmd() = "%02d|%02d|%02d".format(this / 3600, this % 3600 / 60, this % 60)
        if (s.isEmpty()) return ""

        val results = s.split(", ").map { it.toSecs() }.sorted()
        val range = results.last() - results.first()
        val average = results.sum() / results.size
        val median = ((results.size - 1) / 2).let { results.drop(it).dropLast(it).average().toInt() }
        return "Range: ${range.toHmd()} Average: ${average.toHmd()} Median: ${median.toHmd()}"
    }
}

the second is also similar:

package solution

object Stat {
    private fun Int.toHHMMSS(): String =
      "%02d|%02d|%02d".format(this / 3600, this / 60 % 60, this % 60)

    fun stat(s: String): String =
        """(\d{1,2})\|(\d{1,2})\|(\d{1,2})""".toRegex()
            .findAll(s)
            .map { it.groupValues[1].toInt()*3600 + it.groupValues[2].toInt()*60 + it.groupValues[3].toInt() }
            .sorted()
            .toList()
            .run {
                val range = last() - first()
                val average = average().toInt()
                val median =
                    if (size % 2 == 1)
                        get(size / 2)
                    else
                        (get(size / 2) + get(size / 2 - 1)) / 2
                "Range: ${range.toHHMMSS()} Average: ${average.toHHMMSS()} Median: ${median.toHHMMSS()}"
            }
}

+++ src/main/kotlin/Kyu6/Statistics for an Athletic Association.kt
@@ -1,7 +1,5 @@
 package Kyu6
 
-import kotlin.math.round
-
 object Stat {
 
     fun stat(s: String): String {
@@ -13,10 +11,10 @@ object Stat {
         }
         values.forEach { println(it) }
         values.sort()
-        var range=values.last().toInt()-values.first().toInt()
-        var mean=values.sumOf { it.toInt() }/values.size
-        var median=if (values.size%2==0) (values[values.size/2].toInt()+values[values.size/2+1].toInt())/2
-                    else values[(values.size/2).toInt()]
+        var range= values.last() - values.first()
+        var mean=values.sumOf { it }/values.size
+        var median=if (values.size%2==0) (values[values.size/2] + values[values.size/2+1])/2
+                    else values[(values.size/2)]
         return ("Range: "+formatting(range)+" Average: "+ formatting(mean)+" Median: "+ formatting(median))
     }
     fun formatting(seconds:Int):String{
